StripePaymentGateway Integration

+++ src/main/java/com/example/paymentservice/PaymentGateway/IPaymentGateway.java
@@ -1,5 +1,5 @@
 package com.example.paymentservice.PaymentGateway;
 
 public interface IPaymentGateway {
-     String getPaymentLink(int amount, String phoneNumber,String orderId, String customerName);
+     String getPaymentLink(Long amount, String phoneNumber,String orderId, String customerName);
 }
